requirements.txt
streamlit>=1.20.0
numpy>=1.21.0
Pillow>=9.0.0
qrcode>=7.3.1
